torque read added, loop counting, code cleaned up
