+++ platformio.ini
@@ -27,10 +27,4 @@ build_src_filter = +<*> - <.git/> -<.vscode/> -<motor/>
 platform = teensy
 board = teensy40
 framework = arduino
-lib_deps = 
-	sparkfun/SparkFun SerLCD Arduino Library@^1.0.9
-	sparkfun/SparkFun MicroPressure Library@^1.0.1
-	sparkfun/SparkFun MCP9600 Thermocouple Library@^1.0.5
-	br3ttb/PID@^1.2.1
-	thomasfredericks/Bounce2@^2.71
-build_src_filter = +<*> - <.git/> -<.vscode/> -<motor/>
+build_src_filter = +<*> - <.git/> -<.vscode/> -<display/>
